+++ packages/gamut-system/src/types/system.ts
@@ -86,9 +86,14 @@ export type System<
 > = {
   /** Higher order variant function, with two overloads */
   variant: {
-    /** Customizable prop interface */
-    <Prop extends Readonly<string>, Keys extends string>(config: {
-      prop: Prop;
+    /** Customizable default and prop interface */
+    <
+      Keys extends string,
+      Default extends Keys,
+      Prop extends Readonly<string> = 'variant'
+    >(config: {
+      prop?: Prop;
+      default?: Default;
       variants: Readonly<Record<Keys, AllSystemProps<Theme, Config>>>;
     }): (props: WeakRecord<Prop, Keys> & { theme?: Theme }) => CSSObject;
     /** Default `variant` interface */